+++ src/modules/filters/utf8latin1.cpp
@@ -20,6 +20,7 @@
  *
  */
 
+#include <stdint.h>
 #include <utf8latin1.h>
 #include <swbuf.h>
 
@@ -38,7 +39,7 @@ char UTF8Latin1::processText(SWBuf &text, const SWKey *key, const SWModule *modu
   unsigned long uchar;
   unsigned char significantFirstBits, subsequent;
   
-  if ((unsigned long)key < 2) {// hack, we're en(1)/de(0)ciphering
+  if ((uintptr_t)key < 2) {// hack, we're en(1)/de(0)ciphering
 	return (char)-1;
   }
