+++ src/modules/filters/utf8nfkd.cpp
@@ -23,6 +23,7 @@
 
 #ifdef _ICU_
 
+#include <stdint.h>
 #include <utf8nfkd.h>
 #include <swbuf.h>
 
@@ -56,7 +57,7 @@ char UTF8NFKD::processText(SWBuf &text, const SWKey *key, const SWModule *module
 	UErrorCode err = U_ZERO_ERROR;
 	UChar *source, *target;
 
-	if ((unsigned long)key < 2)	// hack, we're en(1)/de(0)ciphering
+	if ((uintptr_t)key < 2)	// hack, we're en(1)/de(0)ciphering
 		return -1;
         
 	int32_t len =  5 + text.length() * 5;
